// webpack-stream-MUST-BE-UPDATED-FOR-NEW-WEBPACK-VERSION-5/webpack.d.ts
export interface AssetEmittedInfo {
    content: Buffer;
    source: Source;
    compilation: Compilation;
    outputPath: string;
    targetPath: string;
}
// ...
export interface CompilationParams {
    normalModuleFactory: NormalModuleFactory;
    contextModuleFactory: ContextModuleFactory;
}
export class Compiler {
    constructor(context: string);
    hooks: Readonly<{
        initialize: SyncHook<[], void>;
        shouldEmit: SyncBailHook<[ Compilation ], boolean>;
        done: AsyncSeriesHook<[ Stats ]>;
        afterDone: SyncHook<[ Stats ], void>;
        additionalPass: AsyncSeriesHook<[]>;
        beforeRun: AsyncSeriesHook<[ Compiler ]>;
        run: AsyncSeriesHook<[ Compiler ]>;
        emit: AsyncSeriesHook<[ Compilation ]>;
        assetEmitted: AsyncSeriesHook<[ string, AssetEmittedInfo ]>;
        afterEmit: AsyncSeriesHook<[ Compilation ]>;
        thisCompilation: SyncHook<[ Compilation, CompilationParams ], void>;
        compilation: SyncHook<[ Compilation, CompilationParams ], void>;
        normalModuleFactory: SyncHook<[ NormalModuleFactory ], void>;
        contextModuleFactory: SyncHook<[ ContextModuleFactory ], void>;
        beforeCompile: AsyncSeriesHook<[ CompilationParams ]>;
        compile: SyncHook<[ CompilationParams ], void>;
        make: AsyncParallelHook<[ Compilation ]>;
        finishMake: AsyncParallelHook<[ Compilation ]>;
        afterCompile: AsyncSeriesHook<[ Compilation ]>;
        watchRun: AsyncSeriesHook<[ Compiler ]>;
        failed: SyncHook<[ Error ], void>;
        invalid: SyncHook<[ string, number ], void>;
        watchClose: SyncHook<[], void>;
        infrastructureLog: SyncBailHook<[ string, string, any[] ], true>;
        environment: SyncHook<[], void>;
        afterEnvironment: SyncHook<[], void>;
        afterPlugins: SyncHook<[ Compiler ], void>;
        afterResolvers: SyncHook<[ Compiler ], void>;
        entryOption: SyncBailHook<[ string, EntryNormalized ], boolean>;
    }>;
    webpack: typeof exports;
    name: string;
    parentCompilation: Compilation;
    root: Compiler;
    outputPath: string;
    watching: Watching;
    outputFileSystem: OutputFileSystem;
    intermediateFileSystem: InputFileSystem &
        OutputFileSystem &
        IntermediateFileSystemExtras;
    inputFileSystem: InputFileSystem;
    watchFileSystem: WatchFileSystem;
    recordsInputPath: string;
    recordsOutputPath: string;
    records: {};
    managedPaths: Set<string>;
    immutablePaths: Set<string>;
    modifiedFiles: Set<string>;
    removedFiles: Set<string>;
    fileTimestamps: Map<string, FileSystemInfoEntry>;
    contextTimestamps: Map<string, FileSystemInfoEntry>;
    resolverFactory: ResolverFactory;
    infrastructureLogger: any;
    options: WebpackOptionsNormalized;
    context: string;
    requestShortener: RequestShortener;
    cache: Cache;
    compilerPath: string;
    running: boolean;
    idle: boolean;
    watchMode: boolean;
    getCache(name: string): CacheFacade;
    getInfrastructureLogger(name: string | (() => string)): WebpackLogger;
    watch(watchOptions: WatchOptions, handler: CallbackFunction<Stats>): Watching;
    run(callback: CallbackFunction<Stats>): void;
    runAsChild(
        callback: (err?: Error, entries?: Chunk[], compilation?: Compilation) => any
    ): void;
    purgeInputFileSystem(): void;
    emitAssets(compilation: Compilation, callback: CallbackFunction<void>): void;
    emitRecords(callback: CallbackFunction<void>): void;
    readRecords(callback: CallbackFunction<void>): void;
    createChildCompiler(
        compilation: Compilation,
        compilerName: string,
        compilerIndex: number,
        outputOptions: OutputNormalized,
        plugins: WebpackPluginInstance[]
    ): Compiler;
    isChild(): boolean;
    createCompilation(): Compilation;
    newCompilation(params: CompilationParams): Compilation;
    createNormalModuleFactory(): NormalModuleFactory;
    createContextModuleFactory(): ContextModuleFactory;
    newCompilationParams(): {
        normalModuleFactory: NormalModuleFactory;
        contextModuleFactory: ContextModuleFactory;
    };
    compile(callback: CallbackFunction<Compilation>): void;
    close(callback: CallbackFunction<void>): void;
}
// ...
export abstract class ContextModuleFactory extends ModuleFactory {
    hooks: Readonly<{
        beforeResolve: AsyncSeriesWaterfallHook<[ any ]>;
        afterResolve: AsyncSeriesWaterfallHook<[ any ]>;
        contextModuleFiles: SyncWaterfallHook<[ string[] ]>;
        alternatives: Pick<
            AsyncSeriesWaterfallHook<[ any[] ]>,
            'tap' | 'tapAsync' | 'tapPromise' | 'name'
        > &
        FakeHookMarker;
        alternativeRequests: AsyncSeriesWaterfallHook<
            [
                any[],
                {
                    mode: 'sync' | 'eager' | 'weak' | 'async-weak' | 'lazy' | 'lazy-once';
                    recursive: boolean;
                    regExp: RegExp;
                    namespaceObject?: boolean | 'strict';
                    addon?: string;
                    chunkName?: string;
                    include?: RegExp;
                    exclude?: RegExp;
                    groupOptions?: RawChunkGroupOptions;
                    category?: string;
                    /**
                     * exports referenced from modules (won't be mangled)
                     */
                    referencedExports?: string[][];
                    resource: string;
                    resourceQuery?: string;
                    resourceFragment?: string;
                    resolveOptions: any;
                }
            ]
        >;
    }>;
    resolverFactory: ResolverFactory;
    resolveDependencies(
        fs: InputFileSystem,
        options: {
            mode: 'sync' | 'eager' | 'weak' | 'async-weak' | 'lazy' | 'lazy-once';
            recursive: boolean;
            regExp: RegExp;
            namespaceObject?: boolean | 'strict';
            addon?: string;
            chunkName?: string;
            include?: RegExp;
            exclude?: RegExp;
            groupOptions?: RawChunkGroupOptions;
            category?: string;
            /**
             * exports referenced from modules (won't be mangled)
             */
            referencedExports?: string[][];
            resource: string;
            resourceQuery?: string;
            resourceFragment?: string;
            resolveOptions: any;
        },
        callback: (err?: Error, dependencies?: ContextElementDependency[]) => any
    ): void;
}
// ...
type EntryNormalized =
    | (() => Promise<EntryStaticNormalized>)
    | EntryStaticNormalized;
// ...
export interface FileSystemInfoEntry {
    safeTime: number;
    timestamp?: number;
    timestampHash?: string;
}
// ...
export interface InputFileSystem {
    readFile: (
        arg0: string,
        arg1: (arg0: NodeJS.ErrnoException, arg1: Buffer) => void
    ) => void;
    readJson?: (
        arg0: string,
        arg1: (arg0: Error | NodeJS.ErrnoException, arg1?: any) => void
    ) => void;
    readlink: (
        arg0: string,
        arg1: (arg0: NodeJS.ErrnoException, arg1: string | Buffer) => void
    ) => void;
    readdir: (
        arg0: string,
        arg1: (arg0: NodeJS.ErrnoException, arg1: string[]) => void
    ) => void;
    stat: (
        arg0: string,
        arg1: (arg0: NodeJS.ErrnoException, arg1: FsStats) => void
    ) => void;
    realpath?: (
        arg0: string,
        arg1: (arg0: NodeJS.ErrnoException, arg1: string) => void
    ) => void;
    purge?: (arg0: string) => void;
    join?: (arg0: string, arg1: string) => string;
    relative?: (arg0: string, arg1: string) => string;
    dirname?: (arg0: string) => string;
}
export interface IntermediateFileSystemExtras {
    mkdirSync: (arg0: string) => void;
    createWriteStream: (arg0: string) => WriteStream;
    rename: (
        arg0: string,
        arg1: string,
        arg2: (arg0: NodeJS.ErrnoException) => void
    ) => void;
}
// ...
export abstract class NormalModuleFactory extends ModuleFactory {
    hooks: Readonly<{
        resolve: AsyncSeriesBailHook<[ ResolveData ], any>;
        resolveForScheme: HookMap<
            AsyncSeriesBailHook<
                [
                    {
                        resource: string;
                        path: string;
                        query: string;
                        fragment: string;
                        data: Record<string, any>;
                    },
                    ResolveData
                ],
                true | void
            >
        >;
        factorize: AsyncSeriesBailHook<[ ResolveData ], any>;
        beforeResolve: AsyncSeriesBailHook<[ ResolveData ], any>;
        afterResolve: AsyncSeriesBailHook<[ ResolveData ], any>;
        createModule: AsyncSeriesBailHook<[ any, ResolveData ], any>;
        module: SyncWaterfallHook<[ Module, any, ResolveData ]>;
        createParser: HookMap<SyncBailHook<any, any>>;
        parser: HookMap<SyncHook<any, void>>;
        createGenerator: HookMap<SyncBailHook<any, any>>;
        generator: HookMap<SyncHook<any, void>>;
    }>;
    resolverFactory: ResolverFactory;
    ruleSet: RuleSet;
    unsafeCache: boolean;
    cachePredicate: Function;
    context: string;
    fs: InputFileSystem;
    parserCache: Map<string, WeakMap<any, any>>;
    generatorCache: Map<string, WeakMap<any, Generator>>;
    resolveResource(
        contextInfo?: any,
        context?: any,
        unresolvedResource?: any,
        resolver?: any,
        resolveContext?: any,
        callback?: any
    ): void;
    resolveRequestArray(
        contextInfo?: any,
        context?: any,
        array?: any,
        resolver?: any,
        resolveContext?: any,
        callback?: any
    ): any;
    getParser(type?: any, parserOptions?: {}): any;
    createParser(type: string, parserOptions?: { [ index: string ]: any; }): Parser;
    getGenerator(type?: any, generatorOptions?: {}): Generator;
    createGenerator(type?: any, generatorOptions?: {}): any;
    getResolver(type?: any, resolveOptions?: any): Resolver & WithOptions;
}
// ...
export interface OutputFileSystem {
    writeFile: (
        arg0: string,
        arg1: string | Buffer,
        arg2: (arg0: NodeJS.ErrnoException) => void
    ) => void;
    mkdir: (arg0: string, arg1: (arg0: NodeJS.ErrnoException) => void) => void;
    stat: (
        arg0: string,
        arg1: (arg0: NodeJS.ErrnoException, arg1: FsStats) => void
    ) => void;
    readFile: (
        arg0: string,
        arg1: (arg0: NodeJS.ErrnoException, arg1: Buffer) => void
    ) => void;
    join?: (arg0: string, arg1: string) => string;
    relative?: (arg0: string, arg1: string) => string;
    dirname?: (arg0: string) => string;
}
// ...
export abstract class RequestShortener {
    contextify: (arg0: string) => string;
    shorten(request: string): string;
}
// ...
export class Stats {
    constructor(compilation: Compilation);
    compilation: Compilation;
    readonly hash: string;
    readonly startTime: any;
    readonly endTime: any;
    hasWarnings(): boolean;
    hasErrors(): boolean;
    toJson(options?: any): any;
    toString(options?: any): string;
}
// ...
export interface WatchFileSystem {
    watch: (
        files: Iterable<string>,
        directories: Iterable<string>,
        missing: Iterable<string>,
        startTime: number,
        options: WatchOptions,
        callback: (
            arg0: Error,
            arg1: Map<string, FileSystemInfoEntry>,
            arg2: Map<string, FileSystemInfoEntry>,
            arg3: Set<string>,
            arg4: Set<string>
        ) => void,
        callbackUndelayed: (arg0: string, arg1: number) => void
    ) => Watcher;
}
// ...
export namespace exports {
    export const webpack: {
        (options: Configuration, callback?: CallbackWebpack<Stats>): Compiler;
        (
            options: Configuration[],
            callback?: CallbackWebpack<MultiStats>
        ): MultiCompiler;
    };
// ...
    export namespace cache {
        export { MemoryCachePlugin };
    }
// ...
}